removed plain password from cache and increased security on registration, removed unique from summoner_name and checking is processed by summoner_name and server simultaneously

+++ src/app/security/guards/jwt_register.guard.ts
@@ -1,34 +1,54 @@
 import * as jwt from 'jsonwebtoken';
-import { ExecutionContext, HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { BadRequestException, ExecutionContext, Injectable, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserRegisterCache } from 'src/database/entity/user_register_cache.entity';
 import { Repository } from 'typeorm';
+import { AuthService } from 'src/app/services/core/auth/auth.service';
 
 @Injectable()
 export class JwtRegisterAuthGuard {
   constructor(
-    @InjectRepository(UserRegisterCache)
-    private readonly userRegisterCacheRepo: Repository<UserRegisterCache>,
+    @InjectRepository(UserRegisterCache) private readonly userRegisterCacheRepo: Repository<UserRegisterCache>,
+    private readonly authService: AuthService,
   ) {}
 
-  canActivate(context: ExecutionContext) {
+  async canActivate(context: ExecutionContext) {
     const request = context.switchToHttp().getRequest();
 
     // validate parameters
     ['id', 'secret'].forEach((param) => {
       if (!request.query[param]) {
-        throw new HttpException(`query parameter {${param}} is missing`, HttpStatus.BAD_REQUEST);
+        throw new BadRequestException(`query parameter {${param}} is missing`);
       }
     });
 
     const { id, secret } = request.query;
 
+    console.log('===');
+    console.log(id);
+    console.log(secret);
+
+    // first check if in register cache
+    const cachedData = await this.authService.retrieveCachedData(id);
+
+    // if secret and cached secret is not exactly same
+    if (secret !== cachedData.secret_token) {
+      throw new UnauthorizedException('Invalid token');
+    }
+
+    console.log('===');
+    console.log(cachedData.id);
+    console.log(cachedData.secret_token);
+    console.log('===');
+
     jwt.verify(secret, process.env.JWT_REGISTER_CACHE_SECRET, (err: jwt.VerifyErrors) => {
       if (err) {
+        console.log(err);
+
         // clear cache with that id if err found on token
         this.userRegisterCacheRepo.delete(id);
 
-        throw new HttpException(err, HttpStatus.UNAUTHORIZED);
+        throw new UnauthorizedException(err);
       }
     });
